backend/requirements.txt
asgiref==3.11.0
click==8.3.1
colorama==0.4.6
Django==4.1
django-cors-headers==4.5.0
django-filter==23.5
django-unfold==0.50.0
djangorestframework==3.15.1
djangorestframework-gis==1.1
h11==0.16.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.38.0
drf-spectacular
